LayoutTests/platform/ios/fast/events/touch/gesture-event-basic-expected.txt

This tests initializing GestureEvents

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL !('GestureEvent' in window) should be true. Was false.

EVENT: gesturestart
PASS event.type is type
PASS event.bubbles is canBubble
PASS event.cancelable is cancelable
PASS event.view is view
PASS event.detail is detail
PASS event.ctrlKey is ctrlKey
PASS event.altKey is altKey
PASS event.shiftKey is shiftKey
PASS event.metaKey is metaKey
PASS event.target is target
PASS event.scale is scale
PASS event.rotation is rotation
FAIL event.screenX should be undefined (of type undefined). Was 2 (of type number).
FAIL event.screenY should be undefined (of type undefined). Was 3 (of type number).
FAIL event.clientX should be undefined (of type undefined). Was 4 (of type number).
FAIL event.clientY should be undefined (of type undefined). Was 5 (of type number).

EVENT: gesturechange
PASS event.type is type
PASS event.bubbles is canBubble
PASS event.cancelable is cancelable
PASS event.view is view
PASS event.detail is detail
PASS event.ctrlKey is ctrlKey
PASS event.altKey is altKey
PASS event.shiftKey is shiftKey
PASS event.metaKey is metaKey
PASS event.target is target
PASS event.scale is scale
PASS event.rotation is rotation
FAIL event.screenX should be undefined (of type undefined). Was 2 (of type number).
FAIL event.screenY should be undefined (of type undefined). Was 3 (of type number).
FAIL event.clientX should be undefined (of type undefined). Was 4 (of type number).
FAIL event.clientY should be undefined (of type undefined). Was 5 (of type number).

EVENT: gestureend
PASS event.type is type
PASS event.bubbles is canBubble
PASS event.cancelable is cancelable
PASS event.view is view
PASS event.detail is detail
PASS event.ctrlKey is ctrlKey
PASS event.altKey is altKey
PASS event.shiftKey is shiftKey
PASS event.metaKey is metaKey
PASS event.target is target
PASS event.scale is scale
PASS event.rotation is rotation
FAIL event.screenX should be undefined (of type undefined). Was 2 (of type number).
FAIL event.screenY should be undefined (of type undefined). Was 3 (of type number).
FAIL event.clientX should be undefined (of type undefined). Was 4 (of type number).
FAIL event.clientY should be undefined (of type undefined). Was 5 (of type number).
PASS successfullyParsed is true

TEST COMPLETE
